+++ src/plantuml_gui/static/sequence-operations.js
@@ -227,11 +227,12 @@ function sequenceEventListeners() {
     messageEventListeners();
     messageOperationEventListeners();
     noteOperationEventListeners();
+    activationEventListeners();
 }
 
 // Called on every render when diagram type is sequence
 async function setHandlersForSequenceDiagram(pumlcontent, element) {
-    fetchSvgFromPlantUml().then((svgContent) => {
+    fetchSvgFromPlantUml().then(async (svgContent) => {
         element.innerHTML = svgContent;
         const svgContainer = element.querySelector('svg');
         const svg = element.querySelector('g');
@@ -241,10 +242,12 @@ async function setHandlersForSequenceDiagram(pumlcontent, element) {
         }
 
         extractLifelinePositions();
+        await fetchMessagePositions();
         cancelMessageAddMode();
+        cancelActivationAddMode();
 
         handleContextMenuBackground(svgContainer);
-        setupLifelineInteraction(svgContainer);
+        setupLifelineInteraction();
         setupParticipantHandlers(svg.querySelectorAll('*'), svg, element);
         setupMessageHandlers(svg.querySelectorAll('*'), svg);
         setupNoteHandlers(svg.querySelectorAll('*'));